requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
qrcode>=7.4.2
django-crispy-forms>=2.0
python-dotenv>=1.0.0
django-bootstrap4>=23.1